make that a bit more variant specific
